requirements.txt
babel==2.17.0
certifi==2025.4.26
charset-normalizer==3.4.2
courlan==1.3.2
dateparser==1.2.1
faiss-cpu==1.11.0
htmldate==1.9.3
idna==3.10
justext==3.0.2
lxml==5.4.0
lxml-html-clean==0.4.2
numpy==2.2.5
packaging==25.0
pymupdf==1.25.5
pymupdf4llm==0.0.24
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
tld==0.13
trafilatura==2.0.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0